plan: pick `generatePhysicalPlans` to one file.
